finished refactoring for readability
